Create the masked_register_translation module
The skimage.feature.masked_register_translation module is now self-contained. It has the masked normalized cross-correlation function mnxc and the masked_register_translation function.
